Move snippet files
